refactor: enrich MCP geospatial tools metadata and LLM guidance
- add shared read-only/open-world annotations across geospatial and WFS tools
- extract reusable zod input schemas and rely on default MCP success responses
- add dedicated tool-level test suites and wire Jest to a test-specific tsconfig
- localize WFS and altitude error messages
- rewrite the WFS CQL filter resource for LLM-first MCP usage
