+++ src/tools/AssietteSupTool.ts
@@ -1,12 +1,24 @@
 import { MCPTool } from "mcp-framework";
 import { z } from "zod";
+
 import { getAssiettesServitudes, URBANISME_SOURCE } from "../gpf/urbanisme.js";
 import logger from "../logger.js";
+import { READ_ONLY_OPEN_WORLD_TOOL_ANNOTATIONS } from "./toolAnnotations.js";
 
-interface SupInput {
-  lon: number;
-  lat: number;
-}
+const assietteSupInputSchema = z.object({
+  lon: z
+    .number()
+    .min(-180)
+    .max(180)
+    .describe("La longitude du point."),
+  lat: z
+    .number()
+    .min(-90)
+    .max(90)
+    .describe("La latitude du point."),
+});
+
+type AssietteSupInput = z.infer<typeof assietteSupInputSchema>;
 
 const assietteSupResultSchema = z
   .object({
@@ -21,50 +33,20 @@ const assietteSupOutputSchema = z.object({
   results: z.array(assietteSupResultSchema).describe("La liste des assiettes de servitudes d'utilité publique pertinentes pour le point demandé."),
 });
 
-class AssietteSupTool extends MCPTool<SupInput> {
+class AssietteSupTool extends MCPTool<AssietteSupInput> {
   name = "assiette_sup";
   title = "Servitudes d’utilité publique";
+  annotations = READ_ONLY_OPEN_WORLD_TOOL_ANNOTATIONS;
   description = `Renvoie, pour un point donné par sa longitude et sa latitude, la liste des assiettes de servitudes d'utilité publique (SUP) pertinentes à proximité, avec leurs propriétés associées. Les résultats peuvent inclure des assiettes ponctuelles, linéaires ou surfaciques. (source : ${URBANISME_SOURCE}).`;
   protected outputSchemaShape = assietteSupOutputSchema;
 
-  schema = z.object({
-    lon: z
-      .number()
-      .min(-180)
-      .max(180)
-      .describe("La longitude du point."),
-    lat: z
-      .number()
-      .min(-90)
-      .max(90)
-      .describe("La latitude du point."),
-  });
+  schema = assietteSupInputSchema;
 
-  async execute(input: SupInput) {
+  async execute(input: AssietteSupInput) {
     logger.info(`assiette_sup(${input.lon},${input.lat})...`);
     return {
       results: await getAssiettesServitudes(input.lon, input.lat),
     };
-  }
-
-  protected createSuccessResponse(data: unknown) {
-    if (
-      typeof data === "object" &&
-      data !== null &&
-      "results" in data &&
-      Array.isArray(data.results)
-    ) {
-      return {
-        content: [
-          {
-            type: "text" as const,
-            text: JSON.stringify(data.results),
-          },
-        ],
-      };
-    }
-
-    return super.createSuccessResponse(data);
   }
 }
