+++ src/tools/CadastreTool.ts
@@ -1,7 +1,24 @@
 import { MCPTool } from "mcp-framework";
 import { z } from "zod";
+
 import { getParcellaireExpress, PARCELLAIRE_EXPRESS_TYPES, PARCELLAIRE_EXPRESS_SOURCE } from "../gpf/parcellaire-express.js";
 import logger from "../logger.js";
+import { READ_ONLY_OPEN_WORLD_TOOL_ANNOTATIONS } from "./toolAnnotations.js";
+
+const cadastreInputSchema = z.object({
+  lon: z
+    .number()
+    .min(-180)
+    .max(180)
+    .describe("La longitude du point."),
+  lat: z
+    .number()
+    .min(-90)
+    .max(90)
+    .describe("La latitude du point."),
+});
+
+type CadastreInput = z.infer<typeof cadastreInputSchema>;
 
 const cadastreResultSchema = z
   .object({
@@ -17,55 +34,20 @@ const cadastreOutputSchema = z.object({
   results: z.array(cadastreResultSchema).describe("La liste des objets cadastraux les plus proches du point demandé."),
 });
 
-interface CadastreInput {
-  lon: number;
-  lat: number;
-}
-
 class CadastreTool extends MCPTool<CadastreInput> {
   name = "cadastre";
   title = "Informations cadastrales";
+  annotations = READ_ONLY_OPEN_WORLD_TOOL_ANNOTATIONS;
   description = `Renvoie, pour un point donné par sa longitude et sa latitude, la liste des objets cadastraux (${PARCELLAIRE_EXPRESS_TYPES.join(', ')}) les plus proches, avec leurs informations associées. Les résultats sont retournés au plus une fois par type lorsqu'ils sont disponibles. (source : ${PARCELLAIRE_EXPRESS_SOURCE}).`;
   protected outputSchemaShape = cadastreOutputSchema;
 
-  schema = z.object({
-    lon: z
-      .number()
-      .min(-180)
-      .max(180)
-      .describe("La longitude du point."),
-    lat: z
-      .number()
-      .min(-90)
-      .max(90)
-      .describe("La latitude du point."),
-  });
+  schema = cadastreInputSchema;
 
   async execute(input: CadastreInput) {
     logger.info(`cadastre(${input.lon},${input.lat})...`);
     return {
       results: await getParcellaireExpress(input.lon, input.lat),
     };
-  }
-
-  protected createSuccessResponse(data: unknown) {
-    if (
-      typeof data === "object" &&
-      data !== null &&
-      "results" in data &&
-      Array.isArray(data.results)
-    ) {
-      return {
-        content: [
-          {
-            type: "text" as const,
-            text: JSON.stringify(data.results),
-          },
-        ],
-      };
-    }
-
-    return super.createSuccessResponse(data);
   }
 }
