+++ src/tools/GpfWfsListTypesTool.ts
@@ -1,20 +1,25 @@
 import { MCPTool } from "mcp-framework";
+import { z } from "zod";
 
+import { READ_ONLY_OPEN_WORLD_TOOL_ANNOTATIONS } from "./toolAnnotations.js";
 import { wfsClient } from "../gpf/wfs.js";
 
-interface WfsTypesInput {}
+const gpfWfsListTypesInputSchema = z.object({});
 
-class GpfWfsListTypesTools extends MCPTool<WfsTypesInput> {
+type GpfWfsListTypesInput = z.infer<typeof gpfWfsListTypesInputSchema>;
+
+class GpfWfsListTypesTool extends MCPTool<GpfWfsListTypesInput> {
   name = "gpf_wfs_list_types";
   title = "Liste complète des types WFS";
+  annotations = READ_ONLY_OPEN_WORLD_TOOL_ANNOTATIONS;
   description = [
     "Renvoie la liste complète des types WFS de la Géoplateforme (GPF).",
     "Utiliser ce tool pour un inventaire exhaustif ou une exploration globale du catalogue.",
     "Pour trouver rapidement un type pertinent à partir de mots-clés, utiliser de préférence gpf_wfs_search_types.",
   ].join("\r\n");
-  schema = {};
+  schema = gpfWfsListTypesInputSchema;
 
-  async execute(input: WfsTypesInput) {
+  async execute(input: GpfWfsListTypesInput) {
     const featureTypes = await wfsClient.getFeatureTypes();
     return featureTypes.map((featureType) => ({
       id: featureType.id,
@@ -24,4 +29,4 @@ class GpfWfsListTypesTools extends MCPTool<WfsTypesInput> {
   }
 }
 
-export default GpfWfsListTypesTools;
+export default GpfWfsListTypesTool;
