+++ src/tools/toolAnnotations.ts
@@ -1,0 +1,6 @@
+export const READ_ONLY_OPEN_WORLD_TOOL_ANNOTATIONS = {
+  readOnlyHint: true,
+  destructiveHint: false,
+  idempotentHint: true,
+  openWorldHint: true,
+} as const;
