+++ test/gpf/parcellaire-express.test.ts
@@ -1,4 +1,3 @@
-import CadastreTool from "../../src/tools/CadastreTool";
 import {getParcellaireExpress} from "../../src/gpf/parcellaire-express.js";
 import { mairieLoray } from "../samples";
 
@@ -50,7 +49,7 @@ const parcellaireExpressFeatureCollection = {
 };
 
 describe("Test getParcellaireExpress",() => {
-    it("should expected values for Mairie de Loray", async () => {
+    it("should return the expected cadastral objects for Mairie de Loray", async () => {
         const c = mairieLoray.coordinates;
         const items : any[] = await getParcellaireExpress(c[0],c[1], async () => parcellaireExpressFeatureCollection);
 
@@ -68,99 +67,5 @@ describe("Test getParcellaireExpress",() => {
             expect(parcelle.idu).toEqual('25349000AD0023');
         }
 
-    });
-});
-
-describe("Test CadastreTool",() => {
-    class TestableCadastreTool extends CadastreTool {
-        async execute() {
-            return {
-                results: await getParcellaireExpress(
-                    mairieLoray.coordinates[0],
-                    mairieLoray.coordinates[1],
-                    async () => parcellaireExpressFeatureCollection
-                ),
-            };
-        }
-    }
-
-    it("should expose an enriched MCP definition", () => {
-        const tool = new CadastreTool();
-        expect(tool.toolDefinition.title).toEqual("Informations cadastrales");
-        expect(tool.toolDefinition.inputSchema.properties?.lon).toMatchObject({
-            type: "number",
-            minimum: -180,
-            maximum: 180,
-        });
-        expect(tool.toolDefinition.inputSchema.properties?.lat).toMatchObject({
-            type: "number",
-            minimum: -90,
-            maximum: 90,
-        });
-        expect(tool.toolDefinition.outputSchema).toBeDefined();
-    });
-
-    it("should return both text content and structuredContent", async () => {
-        const tool = new TestableCadastreTool();
-        const response = await tool.toolCall({
-            params: {
-                name: "cadastre",
-                arguments: {
-                    lon: mairieLoray.coordinates[0],
-                    lat: mairieLoray.coordinates[1],
-                },
-            },
-        });
-
-        expect(response.isError).toBeUndefined();
-        expect(response.content[0]).toMatchObject({
-            type: "text",
-        });
-        const textContent = response.content[0];
-        if (textContent.type !== "text") {
-            throw new Error("expected text content");
-        }
-        const results = JSON.parse(textContent.text);
-        expect(results).toEqual(expect.arrayContaining([
-            expect.objectContaining({
-                type: "parcelle",
-                idu: "25349000AD0023",
-            }),
-        ]));
-        expect(response.structuredContent).toMatchObject({
-            results: expect.arrayContaining([
-                expect.objectContaining({
-                    type: "commune",
-                    nom_officiel: "Loray",
-                }),
-                expect.objectContaining({
-                    type: "parcelle",
-                    idu: "25349000AD0023",
-                }),
-            ]),
-        });
-    });
-
-    it("should return isError=true for invalid input", async () => {
-        const tool = new CadastreTool();
-        const response = await tool.toolCall({
-            params: {
-                name: "cadastre",
-                arguments: {
-                    lon: 600,
-                    lat: 600,
-                },
-            },
-        });
-
-        expect(response.isError).toBe(true);
-        expect(response.content[0]).toMatchObject({
-            type: "text",
-        });
-        const textContent = response.content[0];
-        if (textContent.type !== "text") {
-            throw new Error("expected text content");
-        }
-        expect(textContent.text).toContain("Number must be less than or equal to 180");
     });
 });
