+++ test/gpf/urbanisme.test.ts
@@ -1,5 +1,3 @@
-import AssietteSupTool from "../../src/tools/AssietteSupTool";
-import UrbanismeTool from "../../src/tools/UrbanismeTool";
 import {getUrbanisme, getAssiettesServitudes} from "../../src/gpf/urbanisme.js";
 import { chamonix, mairieLoray } from "../samples";
 
@@ -83,7 +81,7 @@ const assiettesFeatureCollection: {
 };
 
 describe("Test getUrbanisme",() => {
-    it("should expected values for Chamonix", async () => {
+    it("should return the expected urbanisme objects for Chamonix", async () => {
         const c = chamonix.coordinates;
         const items : any[] = await getUrbanisme(c[0],c[1], async () => urbanismeFeatureCollection);
 
@@ -116,7 +114,7 @@ describe("Test getUrbanisme",() => {
 });
 
 describe("Test getAssiettesServitudes",() => {
-    it("should expected values for Loray", async () => {
+    it("should return the expected assiettes for Loray", async () => {
         const c = mairieLoray.coordinates;
         const items : any[] = await getAssiettesServitudes(c[0],c[1], async () => assiettesFeatureCollection);
 
@@ -126,193 +124,5 @@ describe("Test getAssiettesServitudes",() => {
         const names = items.map((item)=>item.nomsuplitt);
         expect(names).toContain("Croix de l'ancien cimetière");
         expect(names).toContain('Fontaine-lavoir');
-    });
-});
-
-describe("Test UrbanismeTool",() => {
-    class TestableUrbanismeTool extends UrbanismeTool {
-        async execute() {
-            return {
-                results: await getUrbanisme(
-                    chamonix.coordinates[0],
-                    chamonix.coordinates[1],
-                    async () => urbanismeFeatureCollection
-                ),
-            };
-        }
-    }
-
-    it("should expose an enriched MCP definition", () => {
-        const tool = new UrbanismeTool();
-        expect(tool.toolDefinition.title).toEqual("Informations d’urbanisme");
-        expect(tool.toolDefinition.inputSchema.properties?.lon).toMatchObject({
-            type: "number",
-            minimum: -180,
-            maximum: 180,
-        });
-        expect(tool.toolDefinition.inputSchema.properties?.lat).toMatchObject({
-            type: "number",
-            minimum: -90,
-            maximum: 90,
-        });
-        expect(tool.toolDefinition.outputSchema).toBeDefined();
-    });
-
-    it("should return both text content and structuredContent", async () => {
-        const tool = new TestableUrbanismeTool();
-        const response = await tool.toolCall({
-            params: {
-                name: "urbanisme",
-                arguments: {
-                    lon: chamonix.coordinates[0],
-                    lat: chamonix.coordinates[1],
-                },
-            },
-        });
-
-        expect(response.isError).toBeUndefined();
-        expect(response.content[0]).toMatchObject({
-            type: "text",
-        });
-        const textContent = response.content[0];
-        if (textContent.type !== "text") {
-            throw new Error("expected text content");
-        }
-        const results = JSON.parse(textContent.text);
-        expect(results).toEqual(expect.arrayContaining([
-            expect.objectContaining({
-                type: "document",
-                du_type: "PLU",
-            }),
-        ]));
-        expect(response.structuredContent).toMatchObject({
-            results: expect.arrayContaining([
-                expect.objectContaining({
-                    type: "document",
-                    du_type: "PLU",
-                }),
-                expect.objectContaining({
-                    type: "zone_urba",
-                    libelle: "Zone U",
-                }),
-            ]),
-        });
-    });
-
-    it("should return isError=true for invalid input", async () => {
-        const tool = new UrbanismeTool();
-        const response = await tool.toolCall({
-            params: {
-                name: "urbanisme",
-                arguments: {
-                    lon: 600,
-                    lat: 600,
-                },
-            },
-        });
-
-        expect(response.isError).toBe(true);
-        expect(response.content[0]).toMatchObject({
-            type: "text",
-        });
-        const textContent = response.content[0];
-        if (textContent.type !== "text") {
-            throw new Error("expected text content");
-        }
-        expect(textContent.text).toContain("Number must be less than or equal to 180");
-    });
-});
-
-describe("Test AssietteSupTool",() => {
-    class TestableAssietteSupTool extends AssietteSupTool {
-        async execute() {
-            return {
-                results: await getAssiettesServitudes(
-                    mairieLoray.coordinates[0],
-                    mairieLoray.coordinates[1],
-                    async () => assiettesFeatureCollection
-                ),
-            };
-        }
-    }
-
-    it("should expose an enriched MCP definition", () => {
-        const tool = new AssietteSupTool();
-        expect(tool.toolDefinition.title).toEqual("Servitudes d’utilité publique");
-        expect(tool.toolDefinition.inputSchema.properties?.lon).toMatchObject({
-            type: "number",
-            minimum: -180,
-            maximum: 180,
-        });
-        expect(tool.toolDefinition.inputSchema.properties?.lat).toMatchObject({
-            type: "number",
-            minimum: -90,
-            maximum: 90,
-        });
-        expect(tool.toolDefinition.outputSchema).toBeDefined();
-    });
-
-    it("should return both text content and structuredContent", async () => {
-        const tool = new TestableAssietteSupTool();
-        const response = await tool.toolCall({
-            params: {
-                name: "assiette_sup",
-                arguments: {
-                    lon: mairieLoray.coordinates[0],
-                    lat: mairieLoray.coordinates[1],
-                },
-            },
-        });
-
-        expect(response.isError).toBeUndefined();
-        expect(response.content[0]).toMatchObject({
-            type: "text",
-        });
-        const textContent = response.content[0];
-        if (textContent.type !== "text") {
-            throw new Error("expected text content");
-        }
-        const results = JSON.parse(textContent.text);
-        expect(results).toEqual(expect.arrayContaining([
-            expect.objectContaining({
-                type: "assiette_sup_s",
-                nomsuplitt: "Croix de l'ancien cimetière",
-            }),
-        ]));
-        expect(response.structuredContent).toMatchObject({
-            results: expect.arrayContaining([
-                expect.objectContaining({
-                    type: "assiette_sup_s",
-                    nomsuplitt: "Croix de l'ancien cimetière",
-                }),
-                expect.objectContaining({
-                    type: "assiette_sup_s",
-                    nomsuplitt: "Fontaine-lavoir",
-                }),
-            ]),
-        });
-    });
-
-    it("should return isError=true for invalid input", async () => {
-        const tool = new AssietteSupTool();
-        const response = await tool.toolCall({
-            params: {
-                name: "assiette_sup",
-                arguments: {
-                    lon: 600,
-                    lat: 600,
-                },
-            },
-        });
-
-        expect(response.isError).toBe(true);
-        expect(response.content[0]).toMatchObject({
-            type: "text",
-        });
-        const textContent = response.content[0];
-        if (textContent.type !== "text") {
-            throw new Error("expected text content");
-        }
-        expect(textContent.text).toContain("Number must be less than or equal to 180");
     });
 });
